refactor(manager): extract connection event subscriptions (#33)
Split state-subscriptions.go (329 lines) into 2 files:
- state-subscriptions.go: Connection state subscriptions (SubscribeConnectionStateChange, GetConnectionStateSubscription) + SimState subscriptions (SubscribeSimStateChange, GetSimStateSubscription) + struct implementations
- connection-event-subscriptions.go: Open subscriptions (SubscribeOnOpen, GetOpenSubscription + struct) + Quit subscriptions (SubscribeOnQuit, GetQuitSubscription + struct)

No behavior changes — pure file reorganization.
